Convocatoria - Configuración de solicitudes
HERCULES-1007 - [CON] IU-CSP-0201-014 - Crear/modificar convocatoria - Configuración de solicitudes

+++ sgi-webapp/src/app/module/csp/convocatoria/convocatoria-formulario/convocatoria-plazos-fases/convocatoria-plazos-fases.component.ts
@@ -14,8 +14,10 @@ import { ConvocatoriaPlazosFaseModalComponent, ConvocatoriaPlazosFaseModalCompon
 import { IConvocatoriaFase } from '@core/models/csp/convocatoria-fase';
 import { DialogService } from '@core/services/dialog.service';
 import { marker } from '@biesbjerg/ngx-translate-extract-marker';
+import { SnackBarService } from '@core/services/snack-bar.service';
 
 const MSG_DELETE = marker('csp.convocatoria.fase.listado.borrar');
+const MSG_ERROR = marker('csp.convocatoria.fase.listado.borrar.error');
 
 @Component({
   selector: 'sgi-convocatoria-plazos-fases',
@@ -31,7 +33,7 @@ export class ConvocatoriaPlazosFasesComponent extends FragmentComponent implemen
   totalElementos: number;
   displayedColumns: string[];
   elementosPagina: number[];
-  public disableAddFase = true;
+  disableAddFase = true;
 
   dataSource: MatTableDataSource<StatusWrapper<IConvocatoriaFase>>;
   @ViewChild(MatPaginator, { static: true }) paginator: MatPaginator;
@@ -41,6 +43,7 @@ export class ConvocatoriaPlazosFasesComponent extends FragmentComponent implemen
 
   constructor(
     protected readonly logger: NGXLogger,
+    protected readonly snackBarService: SnackBarService,
     private actionService: ConvocatoriaActionService,
     private matDialog: MatDialog,
     private readonly dialogService: DialogService
@@ -62,7 +65,10 @@ export class ConvocatoriaPlazosFasesComponent extends FragmentComponent implemen
     this.dataSource = new MatTableDataSource<StatusWrapper<IConvocatoriaFase>>();
     this.dataSource.paginator = this.paginator;
     this.dataSource.sort = this.sort;
-    this.disableAddFase = !Boolean(this.actionService.modeloEjecucionId);
+    this.disableAddFase = !Boolean(this.actionService.getDatosGeneralesConvocatoria().modeloEjecucion);
+
+    this.actionService.initializeConfiguracionSolicitud();
+
     this.subscriptions.push(this.formPart.plazosFase$.subscribe(elements => {
       this.dataSource.data = elements;
       this.logger.debug(ConvocatoriaPlazosFasesComponent.name, 'ngOnInit()', 'end');
@@ -75,11 +81,11 @@ export class ConvocatoriaPlazosFasesComponent extends FragmentComponent implemen
    */
   openModalPlazos(plazo?: StatusWrapper<IConvocatoriaFase>): void {
     this.logger.debug(ConvocatoriaPlazosFasesComponent.name, 'openModalPeriodo()', 'start');
-    const datosPlazosFases = {
+    const datosPlazosFases: ConvocatoriaPlazosFaseModalComponentData = {
       plazos: this.dataSource.data,
       plazo: plazo ? plazo.value : {} as IConvocatoriaFase,
-      idModeloEjecucion: this.actionService.modeloEjecucionId
-    } as ConvocatoriaPlazosFaseModalComponentData;
+      idModeloEjecucion: this.actionService.getDatosGeneralesConvocatoria().modeloEjecucion?.id
+    };
 
     const config = {
       width: GLOBAL_CONSTANTS.widthModalCSP,
@@ -98,18 +104,13 @@ export class ConvocatoriaPlazosFasesComponent extends FragmentComponent implemen
             }
             this.formPart.setChanges(true);
           } else {
-            // TODO coger estos datos del back
-
             this.formPart.addPlazosFases(plazosFase);
           }
         }
         this.logger.debug(ConvocatoriaPlazosFasesComponent.name, 'openModalPeriodo()', 'end');
       }
     );
-
-
     this.logger.debug(ConvocatoriaPlazosFasesComponent.name, 'openModalPeriodo()', 'end');
-
   }
 
   /**
@@ -118,17 +119,21 @@ export class ConvocatoriaPlazosFasesComponent extends FragmentComponent implemen
   deleteFase(wrapper: StatusWrapper<IConvocatoriaFase>) {
     this.logger.debug(ConvocatoriaPlazosFasesComponent.name,
       `${this.deleteFase.name}(${wrapper})`, 'start');
-    this.subscriptions.push(
-      this.dialogService.showConfirmation(MSG_DELETE).subscribe(
-        (aceptado: boolean) => {
-          if (aceptado) {
-            this.formPart.deleteFase(wrapper);
+    if (this.actionService.isDelete(wrapper.value)) {
+      this.subscriptions.push(
+        this.dialogService.showConfirmation(MSG_DELETE).subscribe(
+          (aceptado: boolean) => {
+            if (aceptado) {
+              this.formPart.deleteFase(wrapper);
+            }
+            this.logger.debug(ConvocatoriaPlazosFasesComponent.name,
+              `${this.deleteFase.name}(${wrapper})`, 'end');
           }
-          this.logger.debug(ConvocatoriaPlazosFasesComponent.name,
-            `${this.deleteFase.name}(${wrapper})`, 'end');
-        }
-      )
-    );
+        )
+      );
+    } else {
+      this.snackBarService.showError(MSG_ERROR);
+    }
   }
 
   ngOnDestroy(): void {
